Moved spa_index.jsx to Reports.js
