+++ ui/src/app.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import Reports from './spa_index.jsx';
+import Reports from './Reports.js';
 
 function ReportsPage() {
   return (
